pyproject.toml
[project]
name = "auto-scientist"
version = "0.1.0"
description = "Autonomous scientific modelling framework - LLM-driven model discovery, iteration, and refinement"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk",
    "pydantic>=2.0",
    "click",
    "openai",
    "google-genai",
    "anthropic",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
auto-scientist = "auto_scientist.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_scientist"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
